# swagger.json
{
  "openapi": "3.0.2",
  "info": {
    "description": " Zeal App Backend API - Node.js + Express",
    "version": "1.0.0",
    "title": "Zeal Fitness App",
    "termsOfService": "http://swagger.io/terms/",
    "contact": { "email": "riskdynamix@bimass.com" }
  },
  "basePath": "/api",
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development servers"
    },
    {
      "url": "https://zeal-server-d0353a904474.herokuapp.com/api",
      "description": "Heroku server"
    }
  ],
...
  "schemes": ["https", "http"],
  "paths": {
...
    "/auth/forgot-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Forgot Password",
        "description": "This endpoint allows you to reset your password.",
        "operationId": "forgotPassword",
        "produces": ["application/json", "application/xml"],
        "requestBody": {
          "description": "Feilds for Reseting password",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email":{"type": "string","description": "your email here"}
                }
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "successful operation"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/auth/forgot-password-update": {
      "post": {
        "tags": ["Auth"],
        "summary": "forgot password update",
        "description": "This endpoint allows you to change the password",
        "operationId": "forgotPasswordUpdate",
        "produces": ["application/json", "application/xml"],
        "requestBody": {
          "description": "Feilds for Updating Password",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email":{"type": "string","description": "your email here"},
                  "newPassword":{"type": "string","description": "new password here"}
                }
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "successful operation"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/auth/auth/google": {
      "get": {
        "tags": ["Auth"],
        "summary": "login from google",
        "description": "",
        "operationId": "google login",
        "produces": ["application/json", "application/xml"],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": { "description": "Invalid username/password supplied" }
        },
        "security": [{ "BearerAuth": [] }]
      }
    },
...
  "/post/create-post": {
  "post": {
    "tags": ["Post"],
    "summary": "Create new post",
    "description": "This endpoint allows for the creation of new post.",
    "operationId": "createPost",
    "consumes": ["multipart/form-data"],
    "produces": ["application/json"],
    "requestBody": {
      "description": "Fields for creating post",
      "required": true,
      "content": {
        "multipart/form-data": {
          "schema": {
            "type": "object",
            "properties": {
              "files": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "binary"
                },
                "description": "The files to upload."
              },
              "description": {
                "type": "string",
                "description": "Description of the post"
              },
              "postOwner": {
                "type": "string",
                "description": "Post Owner ID",
                "example": "60b6a23d8f1c4e3d1c8b4567"
              },
              "group": {
                "type": "string",
                "description": "Group ID",
                "example": "6669a3448932aa1ea18af76a"
              }
           
            },
            "required": ["description", "postOwner"]
          }
        }
      }
    },
    "responses": {
      "default": {
        "description": "Successful operation"
      }
    },
    "security": [
      {
        "BearerAuth": []
      }
    ]
  }
},
...
    "/badge/create-badge": {
      "post": {
        "tags": ["Badge", "Desktop"],
        "summary": "createBadge",
        "description": "This endpoint allows you to create a Badge",
        "operationId": "createBadge",
        "produces": ["application/json", "application/xml"],
        "requestBody": {
          "description": "Feilds for creating badge",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name":{"type": "string","description": "name of badge"},
                  "description":{"type": "string","description": "description of the badge"},
                  "type":{"type": "string","description": "type of the badge"},
                  "image":{"type": "string","description": "link of the uploaded picture of the badge"}
                }
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "successful operation"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/badge/get-all-badge": {
      "get": {
        "tags": ["Badge", "Desktop"],
        "summary": "get all badge",
        "description": "This endpoint allows to get all badge",
        "operationId": "getAllBadge",
        "produces": ["application/json", "application/xml"],
      
        "responses": {
          "default": {
            "description": "successful operation"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
"/challenge/create-challenge": {
  "post": {
    "tags": ["Challenge", "Desktop"],
    "summary": "createChallenge",
    "description": "This endpoint allows you to create a challenge",
    "operationId": "createChallenge",
    "produces": ["application/json", "application/xml"],
    "requestBody": {
      "description": "Fields for creating a challenge",
      "required": true,
      "content": {
        "multipart/form-data": {
          "schema": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "description": "Name of the challenge"
              },
              "challengeCreator": {
                "type": "string",
                "description": "Name of the challenge creator"
              },
              "image": {
                "type": "string",
                "format": "binary",
                "description": "Image file of the challenge"
              },
              "badgeCriteria": {
                "type": "string",
                "description": "Criteria for earning a badge"
              },
              "challengeStart": {
                "type": "string",
                "format": "date-time",
                "description": "Start date of the challenge"
              },
              "challengeEnd": {
                "type": "string",
                "format": "date-time",
                "description": "End date of the challenge"
              },
              "type": {
                "type": "string",
                "description": "Type of the challenge"
              },
              "isFeatured": {
                "type": "boolean",
                "description": "Is the challenge featured",
                "default": false
              },
              "days": {
                "type": "array",
                "description": "Active days of the challenge",
                "items": {
                  "type": "object",
                  "properties": {
                    "dayName": {
                      "type": "string",
                      "description": "Name of the day"
                    },
                    "isActive": {
                      "type": "boolean",
                      "description": "Is the day active",
                      "default": false
                    }
                  }
                }
              },
              "user": {
                "type": "array",
                "description": "Users participating in the challenge",
                "items": {
                  "type": "string",
                  "description": "User ID",
                  "example": "60b6a23d8f1c4e3d1c8b4567"
                }
              },
              "exercise": {
                "type": "array",
                "description": "Exercises included in the challenge",
                "items": {
                  "type": "object",
                  "properties": {
                    "exerciseId": {
                      "type": "string",
                      "description": "ID of the exercise",
                      "example": "60b6a23d8f1c4e3d1c8b4568"
                    },
                    "reps": {
                      "type": "number",
                      "description": "Number of repetitions"
                    },
                    "sets": {
                      "type": "number",
                      "description": "Number of sets"
                    },
                    "duration": {
                      "type": "number",
                      "description": "Duration of the exercise in seconds"
                    },
                    "type": {
                      "type": "string",
                      "description": "Type of the exercise (e.g., 'reps' or 'time based')"
                    }
                  }
                }
              },
              "badge": {
                "type": "string",
                "description": "ID of the badge",
                "example": "60b6a23d8f1c4e3d1c8b4569"
              },
              "group": {
                "type": "string",
                "description": "ID of the group",
                "example": "6669a3448932aa1ea18af76a"
              }
            }
          }
        }
      }
    },
    "responses": {
      "default": {
        "description": "successful operation"
      }
    },
    "security": [
      {
        "BearerAuth": []
      }
    ]
  }
},
...
  "/challenge/get-all-featured-challenges": {
    "get": {
      "tags": ["Challenge"],
      "summary": "get all featured challenges",
      "description": "This endpoint allows to get all featured challenge",
      "operationId": "getAllFeaturedChallenges",
      "produces": ["application/json", "application/xml"],
    
      "responses": {
        "default": {
          "description": "successful operation"
        }
      },
      "security": [
        {
          "BearerAuth": []
        }
      ]
    }
  },
  "/challenge/get-popular-challenge": {
    "get": {
      "tags": ["Challenge"],
      "summary": "get popular challenges",
      "description": "This endpoint allows to get all popular challenge",
      "operationId": "getAllPopularChallenges",
      "produces": ["application/json", "application/xml"],
    
      "responses": {
        "default": {
          "description": "successful operation"
        }
      },
      "security": [
        {
          "BearerAuth": []
        }
      ]
    }
  },
  "/challenge/get-user-created-challenges": {
    "get": {
      "tags": ["Challenge"],
      "summary": "get user created challenges",
      "description": "This endpoint allows to get user created challenges",
      "operationId": "getUserCreatedChallenges",
      "produces": ["application/json", "application/xml"],
    "parameters": [
        {
          "name": "id",
          "in": "query",
          "description": "Id of the user",
          "required": true,
          "schema": {
            "type": "string",
            "format": "string"
          }
        }
      ],
      "responses": {
        "default": {
          "description": "successful operation"
        }
      },
      "security": [
        {
          "BearerAuth": []
        }
      ]
    }
  },
...
  "/progress/GetUserProgress": {
    "get": {
      "tags": ["Progress"],
      "summary": "Get User Progress",
      "description": "This endpoint allows to get user progress",
      "operationId": "GetUserProgress",
      "produces": ["application/json", "application/xml"],
      "parameters": [
        {
          "name": "id",
          "in": "query",
          "description": "id of the user",
          "required": false,
          "schema": {
            "type": "string",
            "format": "string"
          }
        },
        {
          "name": "challengeId",
          "in": "query",
          "description": "id of the challenge",
          "required": false,
          "schema": {
            "type": "string",
            "format": "string"
          }
        }
      ],
      "responses": {
        "default": {
          "description": "successful operation"
        }
      },
      "security": [
        {
          "BearerAuth": []
        }
      ]
    }
  },
  "/progress/GetUserExerciseLog": {
    "get": {
      "tags": ["Progress"],
      "summary": "Get User Exercise Log",
      "description": "This endpoint allows to get user Exercise Log",
      "operationId": "GetUserExerciseLog",
      "produces": ["application/json", "application/xml"],
      "parameters": [
        {
          "name": "id",
          "in": "query",
          "description": "id of the user",
          "required": false,
          "schema": {
            "type": "string",
            "format": "string"
          }
        },
        {
          "name": "challengeId",
          "in": "query",
          "description": "id of the challenge",
          "required": false,
          "schema": {
            "type": "string",
            "format": "string"
          }
        }
      ],
      "responses": {
        "default": {
          "description": "successful operation"
        }
      },
      "security": [
        {
          "BearerAuth": []
        }
      ]
    }
  },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}